calculatable module
