+++ spec/spec_helper.rb
@@ -2,3 +2,5 @@ require 'simplecov'
 SimpleCov.start
 
 require_relative 'enigma_spec'
+require_relative 'decrypt_runner_spec'
+require_relative 'encrypt_runner_spec'
